requirements.txt
pika
influxdb-client
python-dotenv

